requirements.txt
click>=6.0
Jinja2>=2.7
PyYAML>=3.12
